+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 | Difficulty | Resolved | Total |
 | :--------- | :------- | :---- |
 | Easy       | 21       | 492   |
-| Medium     | 35       | 837   |
+| Medium     | 36       | 837   |
 | Hard       | 0        | 328   |
 
 ## 0005 - Longest palindromic substring
@@ -246,6 +246,10 @@ $$Pow({x}, {n})=Pow\left({x}^2, \dfrac{n}{2}\right) * Pow\left(x, n \mod 2\right
 ## 0073 Set matrix zeros
 
 - Store x, y to be zeros
+
+## 0074 Search a 2D matrix
+
+- Binary search, twice
 
 ## 0083 - Remove duplicates from sorted list